feat(arc-303): create higher order components to check permissions
