+++ src/pages/account/mijn-historiek/index.tsx
@@ -2,8 +2,7 @@ import { Table } from '@meemoo/react-components';
 import { GetServerSideProps, NextPage } from 'next';
 import { useTranslation } from 'next-i18next';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
-import { useCallback, useEffect, useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 import { Column, TableOptions } from 'react-table';
 import { useQueryParams } from 'use-query-params';
 
@@ -19,8 +18,7 @@ import { AccountLayout } from '@account/layouts';
 import { withAuth } from '@auth/wrappers/with-auth';
 import { withI18n } from '@i18n/wrappers';
 import { Loading, PaginationBar, sortingIcons } from '@shared/components';
-import { useHasAllPermission } from '@shared/hooks/has-permission';
-import { toastService } from '@shared/services/toast-service';
+import { withAllRequiredPermissions } from '@shared/hoc/withAllRequeredPermissions';
 import { OrderDirection, Visit, VisitStatus } from '@shared/types';
 import { createPageTitle } from '@shared/utils';
 import { useGetVisits } from '@visits/hooks/get-visits';
@@ -31,20 +29,6 @@ import { VisitorLayout } from 'modules/visitors';
 const AccountMyHistory: NextPage = () => {
 	const { t } = useTranslation();
 	const [filters, setFilters] = useQueryParams(ACCOUNT_HISTORY_QUERY_PARAM_CONFIG);
-	const canManageAccount: boolean | null = useHasAllPermission(Permission.MANAGE_ACCOUNT);
-	const router = useRouter();
-
-	useEffect(() => {
-		if (!canManageAccount) {
-			toastService.notify({
-				title: t('pages/account/mijn-historiek/index___geen-toegang'),
-				description: t(
-					'pages/account/mijn-historiek/index___je-hebt-geen-rechten-om-deze-pagina-te-bekijken'
-				),
-			});
-			router.replace('/');
-		}
-	}, [canManageAccount, router, t]);
 
 	const visits = useGetVisits({
 		searchInput: undefined,
@@ -93,10 +77,6 @@ const AccountMyHistory: NextPage = () => {
 	const renderEmptyMessage = (): string => {
 		return t('pages/account/mijn-historiek/index___geen-historiek');
 	};
-
-	if (!canManageAccount) {
-		return <Loading fullscreen />;
-	}
 
 	return (
 		<VisitorLayout>
@@ -166,4 +146,4 @@ const AccountMyHistory: NextPage = () => {
 
 export const getServerSideProps: GetServerSideProps = withI18n();
 
-export default withAuth(AccountMyHistory);
+export default withAuth(withAllRequiredPermissions(AccountMyHistory, Permission.MANAGE_ACCOUNT));
